Segment 2 - Commit 2
These are jupyter notebooks used to process the CSVs (some of which were saved in the process of running the code because many of these formulas take a long time to run, so saving the output also saves time later). ORDER TO RUN THE NOTEBOOKS: 1- compiling, 2- cleaning, 3- analyzing. The 'modeling' notebook ended up not contributing toward the ultimate model, but it's a record that I did try machine learning models.